Split USBDeviceShim into a separate class file
